wip: creation of subsystem solver as a plugin
We want solvers for subsystems (what determines if
something is a valid match) to be an interface. E.g.,
the first approach I took was a database, but I also
want to try a python (C++ backed) library that creates
a graph. Mostly because that way would be really fun.
