+++ fractale/defaults.py
@@ -1,3 +1,5 @@
 fractale_dir = ".fractale"
 valid_settings = {"sharedfs", "stage"}
 sharedfs = True
+solver_backends = ["database", "graph"]
+solver_backend_default = "database"
