+++ fractale/subsystem/__init__.py
@@ -1,9 +1,9 @@
 import os
 
-from .subsystem import SubsystemRegistry
+from .subsystem import SubsystemSolver
 
 
-def get_subsystem_registry(path):
+def get_subsystem_solver(path, backend="database"):
     """
     Generate a user subsystem registry, where the structure is expected
     to be a set of <cluster>/<subsystem>. For the FractaleStore, this
@@ -13,4 +13,4 @@ def get_subsystem_registry(path):
         raise ValueError(f"Cluster subsystem root {path} does not exist")
 
     # Generate the subsystem registry
-    return SubsystemRegistry(path)
+    return SubsystemSolver(path, backend)
