+++ fractale/subsystem/solver/__init__.py
@@ -1,0 +1,13 @@
+import os
+
+from .database import DatabaseSolver
+
+
+def load_solver(backend, path):
+    """
+    Load the solver backend
+    """
+    if backend == "database":
+        return DatabaseSolver(path)
+
+    raise ValueError(f"Unsupported backend {backend}")
